Terrain Audio Fix
Again, while the audio files are mostly WIP, an existing issue involving problematic detection or lack thereof from different terrain types has now been solved. As well as this, the WagonAudio script has now been altered to be readily usable in any scene rather than just the first. In terms of application in the scene the WagonAudio script should be applied to the base wagon, and for this scene use Road, RockyTerrain and WoodenTerrain as part of the WagonAudio array. As well as this, the WagonAudio prefab placing I found to be most effective is in the middle of the wagon in terms of x and z, while meeting the wheels about halfway in terms of y position (-0.07,0.60178, 0.14 once added as a child to the base PlayerWagon). Finally, in order to detect the path type, each section of path (e.g: bridges, all road objects and the rocky terrain after the bridge) should be tagged appropriately while either one of the front two wheels on the wagon has a box collider pertruding slightly downwards so that collision between it and the terrain can be detected.