Full Audio and Visual merge (IMPORTANT DESCRIPTION)
All audio and visual features have been merged together, however this seems to have tampered with the navigation of the wagon as once the tree has been destroyed (which the script for requires TreeChopAudio to be used, apologies for not clarifying this in previous commits), the wagon carries on along the road and stops shortly after and doesn't cross the bridge like normal. This is the only thing that seems necessary to fix, otherwise the combination of the audio and visual systems seem good to go.