Code for stopping wagon audio and general prefab for Wagon implemented
The Audio specific features required for the PlayerWagon have now been added, making it easier to use in scenes. As well as this, a check for wagon speed has been added for the Bandit Scene to be used for faster snow wagon audio, and a general check has been added so audio playback stops if the Player Nav Mesh is stopped. However, mechanics for how the wagon stop for the Bandit and Town scenes will need to be looked into as they don't seem to adhere to this method while the River scene does.