Audio fixes, and more bandit polishing
Many audio clips have now been fixed to omit the silences at either the beginning or ending of their audio. This was done by instead using them as .wav files, and slight alterations for each one. Other versions of the clips exist in the assets, but if the new ones are satisfactory the others can be disposed of. 

In the bandit scene, more intense audio will now play for when the wagon is going faster, once again done in Audacity over Unity due to Unity's lack of a straightforward tempo/speed feature. To add, horse whinnies for this scenario have been added, and will be picked at random. There is a small workaround in the script created for this which slows down the clip choosing and makes sure that the source is not playing every frame. 

From here, a crashing clip for the audio in the bandit scene will need to be added, as well as points in which the wagon travelling audio will stop in the town and bandit scenes. After this, optimal spacial points can be worked out for certain audio clips, and more audio can be added to flesh out scenes alongside royalty free music. Work will also commence on the repair scene.