Repair Scene and Misc additions
Overall: The Viewbox's "isMoving" variable has helped establish points at which the wagon's audio needs to stop playing. Generally, there is a second start again for moving afterwards, but as of now the audio stops once the wagon stops.

Misc: Ambience has been added to the start menu and ending scenes in order to give more interest to them.

Bandit Scene: A small change to the wind's position, crashing clip was not required for this scene as the wagon does not crash at the finish.

Repair Scene:
- Ambience with light wind has been implemented, making use of the ZOOM audio that was acquired.
- Working clips for the hammer/nail stalls and the iron stall have been added. Audio for the jack store still needs to be decided. 
- The wagon's audio here uses an altered version of Stage 1's Road audio with added rickety noises while the wagon is faulty, and returns to the normal audio once it has been repaired (this relies on a new variable in RepairController).

To do's: Spacing for previous audio could be optimised, as well as the potential inclusion of royalty free music and additional audio to flesh out the scenes.